WIP Update HA custom component
I struggled to import the VeSync classes properly, so I copied the library to the custom compoent's folder and imported from the current directory (".")  This is a WIP and shouldn't be left this way.

A new class (VeSyncWallSwitchHA) was created to represent the wall switch since it doesn't supprort voltage and power information.